% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_effect.R
\name{info_bal_plots}
\alias{info_bal_plots}
\title{Evaluate propensity scores using balance plots.}
\usage{
info_bal_plots(knowledge_var, covariates, prop_score, data)
}
\arguments{
\item{knowledge_var}{A character string representing a binary knowledge variable.}

\item{covariates}{A concatenated character string of covariates that can be expected to have an effect on knowledge.}

\item{data}{A data frame containing aforementioned variables.}

\item{prop_scores}{A character string representing a vector of propensity scores.}
}
\value{
\code{info_bal_plots} returns a list of balance plots.
}
\description{
\code{info_prop_scores} calculates propensity scores to be used as weights in subsequent, counterfactual modeling,
in order to improve balance. The function uses logistic regression to calculate the scores. It is recommended that
the calculated scores are independently evaluated by way of balane plots, here implemented in \code{info_bal_plots}.
}
\details{
The function uses \link[cobalt]{bal.plot} to generate balance plots.
}
\examples{
knowledge_binary <- sample(c(1,0), replace=TRUE, size=100)
education <- sample(c("none","gcse","alevel","university"), replace=TRUE, size=100)
income <- sample(c("Q1","Q2","Q3","Q4"), replace=TRUE, size=100)
prop_scores <- abs(rnorm(100,0,1))
df <- data.frame(knowledge_binary, education, income, prop_scores)
info_bal_plots(knowledge_var = "knowledge_binary",
               covariates = c("income", "education"),
               prop_score ="prop_score",
               data = df)
}
\seealso{
See \code{\link{info_scale}} for how to construct a knowledge scale, and \code{\link{info_prop_scores}} for
how to generate propensity scores.
}
