% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_effect.R
\name{info_effect}
\alias{info_effect}
\title{Calculate information effects.}
\usage{
info_effect(
  outcome,
  knowledge_var,
  covariates,
  prop_weight,
  survey_weight,
  boot_ci = F,
  data
)
}
\arguments{
\item{outcome}{A character string representing the outcome variable that you are looking to estimate an effect on.}

\item{knowledge_var}{A character string representing a binary knowledge variable.}

\item{covariates}{A concatenated character string of covariates.}

\item{prop_weight}{A character string representing a vector containing propensity scores}

\item{survey_weight}{A character string representing a vector containing survey weights.}

\item{boot_ci}{logical, designates whether to calculate bootstrapped confidence intervals.}

\item{data}{A data frame containing aforementioned variables.}
}
\value{
\code{info_effect} returns a list of the following:
    \itemize{
      \item \code{model} The model used to estimate the information effect
      \item \code{actual_proportion} The actual, survey weighted proportion on the \code{outcome} variable
      \item \code{informed_proportion} The informed, survey weighted proportion on the \code{outcome} variable
      \item \code{difference} The difference between \code{actual_proportion} and the \code{informed_proportion}, i.e., the
      information effect.
      \item \code{actual_upr}, \code{actual_lwr}, \code{informed_upr}, \code{informed_lwr} The upper and lower bounds of the
      bootstrapped confidence intervals on \code{actual_proportion} and \code{informed_proportion}.
   }
}
\description{
\code{info_effect} calculates information effects on the basis of survey data with a binary knowledge
variable, propensity scores, and survey weights, while controlling for a set of covariates. It can
also generate bootstrapped confidence intervals.
}
\examples{
# NOTE: This won't work without sample data
info_effect(outcome = "immigration",
    knowledge_var = "knowledge_binary",
    covariates = c("age",
                   "gender",
                   "education",
                   "income",
                   "religion",
                   "ethnicity"),
    prop_weight = "prop_score",
    survey_weight = "survey_wt",
    data = df,
    boot_ci = T)
}
\seealso{
See \code{\link{info_scale}} for how to construct a knowledge scale, \code{\link{info_emmeans}} for evaluating construct
validity for such a scale, and \code{\link{info_prop_scores}} for calculating propensity scores.
}
