% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_effect.R
\name{info_prop_scores}
\alias{info_prop_scores}
\title{Calculate propensity scores for counterfactual modeling.}
\usage{
info_prop_scores(knowledge_var, covariates, data)
}
\arguments{
\item{knowledge_var}{A character string representing a binary knowledge variable.}

\item{covariates}{A concatenated character string of covariates that can be expected to have an effect on knowledge.}

\item{data}{A data frame containing aforementioned variables.}
}
\value{
\code{info_prop_scores} returns a vector of propensity scores.
}
\description{
\code{info_prop_scores} calculates propensity scores to be used as weights in subsequent, counterfactual modeling,
in order to improve balance. The function uses logistic regression to calculate the scores. It is recommended that
the calculated scores are independently evaluated, e.g., using the balance plots in \code{cobalt::bal.plot}.
}
\details{
The function calculates the probability of an observation being found in the 'informed' category, represented by 1 on the
binary knowledge variable, as a function of a set of covariates. Call that probability \code{p}. The scores are then calculated as
the inverse of that probability, e.g., \code{1/p}, for anyone in the 'informed' category, and \code{1/(1-p)} otherwise.
}
\examples{
knowledge_binary <- sample(c(1,0), replace=TRUE, size=100)
education <- sample(c("none","gcse","alevel","university"), replace=TRUE, size=100)
income <- sample(c("Q1","Q2","Q3","Q4"), replace=TRUE, size=100)
df <- data.frame(knowledge_binary, education, income)
info_prop_scores(knowledge_var = "knowledge_binary",
                 covariates = c("education","income"),
                 data = df)
}
\seealso{
See \code{\link{info_scale}} for how to construct a knowledge scale.
}
