% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_effect.R
\name{info_scale}
\alias{info_scale}
\title{Construct a knowledge scale using IRT modeling.}
\usage{
info_scale(items, data, binary_cutoff = 0.9)
}
\arguments{
\item{items}{A concatenated character string of names of the knowledge items.}

\item{data}{A data frame containing a set of binary knowledge items (\code{1} = true, \code{0} = false).}

\item{binary_cutoff}{Percentile value for binary knowledge scale, with respondents
at or above that value coded as \code{1} and otherwise as \code{0}. Set to 90th
percentile by default.}
}
\value{
\code{inf_scale} returns a list of the following:
    \itemize{
      \item \code{model} The IRT model
      \item \code{model_summary} A summary of the IRT model
      \item \code{model_coef} The coefficients for the IRT model
      \item \code{know_scores} A vector of knowledge scores
      \item \code{know_scores_binary} A vector of binary knowledge scores
      \item \code{know_scores_binary_tbl} A table with the proportions on the binary knowledge variable
      \item \code{par_analysis} The plot for a paralell analysis of the knowledge items, to evaluate dimensionality.
      \item \code{q3} The Q3 values for the IRT model, to investigate local independence
      \item \code{empirical_plots} Empirical plots for the IRT model, to evaluate model fit
   }
}
\description{
\code{info_scale} calculates a knowledge scale using Item Response Theory (IRT) modeling
on the basis of a set of binary knowledge items. The function uses the \code{mirt} package
for IRT modeling.
}
\details{
Further diagnostics can be accessed through the \code{model} object returned by the function:
\code{plot(model, type = "trace")} gives the trace plot and \code{plot(model, type = "info")} gives the
information plot.
}
\examples{
item1 <- sample(c(0,1), replace=TRUE, size=100)
item2 <- sample(c(0,1), replace=TRUE, size=100)
item3 <- sample(c(0,1), replace=TRUE, size=100)
df <- data.frame(item1, item2, item3)
irt_model <- info_scale(items = c("item1", "item2","item3"),
                        data = df)
}
