% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_effect.R
\name{info_emmeans}
\alias{info_emmeans}
\title{Evaluate construct validity using marginal means.}
\usage{
info_emmeans(knowledge_var, covariates, data)
}
\arguments{
\item{covariates}{A concatenated character string of covariates across which marginal means will be estimated.}

\item{data}{A data frame containing aforementioned variables.}

\item{knolwedge_var}{A character string representing a knowledge variable.}
}
\value{
\code{info_emmeans} returns a list of marginal means for the covariates in question.
}
\description{
\code{info_emmeans} estimates marginal mean levels of knowledge using the \code{emmeans} package for different demographic
variables in order to evaluate construct validity for the underlying knowledge scale.
}
\details{
For a plausible knowledge scale, we would expect to see a positive relationship between knowledge and education and
income, and for men to know more than women.
}
\examples{
# NOTE: This won't work without sample data
info_emmeans(knowledge_var = "knowledge",
            covariates = c("income", "education"),
            data = df)
}
\seealso{
See \code{\link{info_scale}} for how to construct a knowledge scale.
}
